src/requirements.txt
fastapi~=0.115.6
uvicorn
pymongo
pyjwt
passlib
argon2-cffi
python-multipart